Adjustments for new computer

+++ hellamouse/core.h
@@ -1,10 +1,10 @@
-const byte DEBUG = 1;
-const byte LENGTH = 16;
+const int DEBUG = 1;
+const int LENGTH = 16;
 
-const byte NORTH = 1;
-const byte EAST = 2;
-const byte SOUTH = 4;
-const byte WEST = 8;
-const byte ALLWALLS = 15;
-const byte UNDEFINED = 255; //we won't have a spiral maze. if this val is > 255, better change the flip method
+const int NORTH = 1;
+const int EAST = 2;
+const int SOUTH = 4;
+const int WEST = 8;
+const int ALLWALLS = 15;
+const int UNDEFINED = 255; //we won't have a spiral maze. if this val is > 255, better change the flip method
